requirements.txt
flask
numpy
pandas
pytest
pyhamcrest